pyproject.toml
[project]
name = "lolicore"
version = "0.1.0"
description = "Simple, Playable and Flexible Language Model"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.3.2",
    "flax==0.10.3",
    "jax==0.5.0",
    "optax==0.2.4",
    "orbax-checkpoint>=0.11.6",
    "orbax-export>=0.0.6",
    "tokenizers>=0.21.0",
    "transformers>=4.49.0",
    "wandb>=0.19.7",
    "libtpu>=0.0.10; platform_system != 'Darwin'",
    "matplotlib>=3.10.1",
    "tensorflow>=2.18.0",
    "tf2onnx>=1.16.1",
]
